requirements.txt
langchain
langchain-google-genai
langchain-core
langchain-community
google-genai
groq
pyPDF2
faiss-cpu
streamlit
python-dotenv
streamlit-audiorecorder
# chromadb
# llama-index
# pypdf   